requirements.txt
# # Core dependencies
# pymongo==4.6.0
# redis==5.0.0
# python-dotenv==1.0.0
# pandas==2.2.3
# numpy==1.24.3

# # ML/AI dependencies
# sentence-transformers==2.2.2
# faiss-cpu==1.7.4

# # API dependencies
# fastapi==0.104.1
# uvicorn==0.24.0
# pydantic==2.3.0
# python-multipart==0.0.6

# # Task scheduling
# celery==5.3.2

# # Testing
# pytest==7.4.0
# pytest-cov==4.1.0

pymongo
redis
python-dotenv
pandas
numpy
sentence-transformers
faiss-cpu
fastapi
uvicorn
pydantic
python-multipart
celery
pytest
pytest-cov

upstash-redis

pydantic-core